requirements.txt
tweepy>=4.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
feedparser>=6.0.0
schedule>=1.2.0
rich>=13.0.0 